+++ packages/web3-eth/src/index.ts
@@ -45,6 +45,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 /**
  *
  */
+import 'setimmediate';
 
 import { Web3Eth } from './web3_eth';